+++ src/local_rag_backend/cli_commands/server.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+import click
+import uvicorn
+
+from local_rag_backend.settings import settings
+
+
+@click.command("server")
+def server_cmd() -> None:
+    """Start the RAG FastAPI server using settings from config file or environment."""
+    click.echo(f"🚀 Starting server on {settings.app_host}:{settings.app_port}...")
+    click.echo(f"   - Mode: {'Development (reload)' if settings.debug else 'Production'}")
+    click.echo(f"   - Log level: {settings.log_level.upper()}")
+    click.echo(f"   - Retrieval: {settings.retrieval_mode}")
+
+    uvicorn.run(
+        "local_rag_backend.app.main:app",
+        host=settings.app_host,
+        port=settings.app_port,
+        reload=settings.debug,
+        log_level=settings.log_level.lower(),
+    )
